Change: (select-packages) Add test-fn argument
This helps when dealing with inconsistent data, like author/maintainer
headers that are supposed to be the same person but may not always be
exactly the same string.  By using e.g. string=, we can find packages
with an exact string, helping to find a package that needs its headers
fixed.